Use -1 as an avg sentinel value and fix a bug when deferring waste when a historical item was typed in on the shopping list
The modal would not properly display and execute the ask me later function when invoked from the shopping list. It now passes a value of 'deferred' when the ask me later button is pressed and this is checked to properly move items to shopping when no waste is logged but the item should still be moved.